remove elasticsearch mock (#101)
