+++ case-server/src/main/java/com/powsybl/caseserver/elasticsearch/CaseInfosRepository.java
@@ -7,8 +7,6 @@
 package com.powsybl.caseserver.elasticsearch;
 
 import com.powsybl.caseserver.dto.CaseInfos;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
@@ -18,8 +16,6 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
  *
  * @author Slimane Amar <slimane.amar at rte-france.com>
  */
-@ConditionalOnExpression("'${spring.data.elasticsearch.enabled:false}' == 'true'")
-@Lazy
 public interface CaseInfosRepository extends ElasticsearchRepository<CaseInfos, String> {
 
     Page<CaseInfos> findByUuid(String id, Pageable pageable);
