--- case-server/src/main/java/com/powsybl/caseserver/elasticsearch/CaseInfosServiceImpl.java
@@ -1,85 +1,0 @@
-/**
- * Copyright (c) 2020, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.powsybl.caseserver.elasticsearch;
-
-import com.google.common.collect.Lists;
-import com.powsybl.caseserver.dto.CaseInfos;
-import java.util.List;
-import java.util.Optional;
-import org.elasticsearch.index.query.QueryBuilders;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.elasticsearch.core.ElasticsearchOperations;
-import org.springframework.data.elasticsearch.core.query.NativeSearchQuery;
-import org.springframework.data.elasticsearch.core.query.NativeSearchQueryBuilder;
-import org.springframework.lang.NonNull;
-
-/**
- * A class to implement metadatas transfer in the DB elasticsearch
- *
- * @author Slimane Amar <slimane.amar at rte-france.com>
- */
-@ComponentScan(basePackageClasses = {CaseInfosRepository.class})
-public class CaseInfosServiceImpl implements CaseInfosService {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(CaseInfosServiceImpl.class);
-
-    @Autowired
-    private CaseInfosRepository caseInfosRepository;
-
-    @Autowired
-    private ElasticsearchOperations operations;
-
-    @Override
-    public CaseInfos addCaseInfos(@NonNull final CaseInfos ci) {
-        caseInfosRepository.save(ci);
-        return ci;
-    }
-
-    @Override
-    public Optional<CaseInfos> getCaseInfosByUuid(@NonNull final String uuid) {
-        Page<CaseInfos> res = caseInfosRepository.findByUuid(uuid,  PageRequest.of(0, 1));
-        return res.get().findFirst();
-    }
-
-    @Override
-    public List<CaseInfos> getAllCaseInfos() {
-        return Lists.newArrayList(caseInfosRepository.findAll());
-    }
-
-    @Override
-    public List<CaseInfos> searchCaseInfos(@NonNull final String query) {
-        NativeSearchQuery searchQuery = new NativeSearchQueryBuilder().withQuery(QueryBuilders.queryStringQuery(query)).build();
-        return Lists.newArrayList(operations.search(searchQuery, CaseInfos.class)
-                                            .map(searchHit -> searchHit.getContent()));
-    }
-
-    @Override
-    public void deleteCaseInfos(@NonNull final CaseInfos ci) {
-        caseInfosRepository.delete(ci);
-    }
-
-    @Override
-    public void deleteCaseInfosByUuid(@NonNull  String uuid) {
-        caseInfosRepository.deleteById(uuid);
-    }
-
-    @Override
-    public void deleteAllCaseInfos() {
-        caseInfosRepository.deleteAll(getAllCaseInfos());
-    }
-
-    @Override
-    public void recreateAllCaseInfos(List<CaseInfos> caseInfos) {
-        caseInfosRepository.deleteAll();
-        caseInfosRepository.saveAll(caseInfos);
-    }
-}
