+++ case-server/src/main/java/com/powsybl/caseserver/elasticsearch/ESConfig.java
@@ -11,10 +11,8 @@ import org.joda.time.DateTime;
 import org.joda.time.format.DateTimeFormatter;
 import org.joda.time.format.ISODateTimeFormat;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.data.convert.ReadingConverter;
 import org.springframework.data.convert.WritingConverter;
@@ -39,7 +37,6 @@ import java.util.Optional;
 
 @Configuration
 @EnableElasticsearchRepositories
-@Lazy
 public class ESConfig extends AbstractElasticsearchConfiguration {
 
     @Value("#{'${spring.data.elasticsearch.embedded:false}' ? 'localhost' : '${spring.data.elasticsearch.host}'}")
@@ -56,18 +53,6 @@ public class ESConfig extends AbstractElasticsearchConfiguration {
 
     @Value("${spring.data.elasticsearch.password:#{null}}")
     private Optional<String> password;
-
-    @Bean
-    @ConditionalOnExpression("'${spring.data.elasticsearch.enabled:false}' == 'true'")
-    public CaseInfosService caseInfosServiceImpl() {
-        return new CaseInfosServiceImpl();
-    }
-
-    @Bean
-    @ConditionalOnExpression("'${spring.data.elasticsearch.enabled:false}' == 'false'")
-    public CaseInfosService caseInfosServiceMock() {
-        return new CaseInfosServiceMock();
-    }
 
     @Bean
     @Override
