+++ case-server/src/test/java/com/powsybl/caseserver/CaseControllerTest.java
@@ -57,7 +57,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
  */
 @RunWith(SpringRunner.class)
 @AutoConfigureMockMvc
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK, properties = {"spring.data.elasticsearch.enabled=true"},
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.MOCK, properties = {"test.elasticsearch.enabled=true"},
         classes = { EmbeddedElasticsearch.class, CaseController.class, TestChannelBinderConfiguration.class})
 @TestPropertySource(properties = {"case-store-directory=/cases"})
 public class CaseControllerTest {
